Create boutina.adventure.c
Program is a game that randomly builds a small graph of connected rooms, creating a file for each room with pertinent information. Then reads in that information and starts a text based game where you enter a room to travel to. The goal is to go from start to finish, so it is a small maze. Upon completion the program prints how many rooms you traversed to reach the end.